Chore: PlanPlaceListDetailItem 파일 분리, css수정
